refactor: remove FormColumns.php trait
